api-central/requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.19.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
